Removes dependency on Core and adds ext method to Api instead
